requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
colorama==0.4.6
    # via -r requirements.in
gitdb==4.0.11
    # via gitpython
gitpython==3.1.40
    # via -r requirements.in
smmap==5.0.1
    # via gitdb
